+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Recent and upcoming changes to lightdash
 ## [Unreleased]
 ### Added
  - CSV export feature for table of results
- - app url is shareable and can be used to share your current work (active table, columns, sorts)
+ - app url is shareable and can be used to share your current work (active table, columns, sorts, filters)
  - user is prompted to confirm before losing work (starting to explore a new table)
  - dbt refreshes happen in the background whenever possible